requirements.txt
llama-cpp-python
langchain
faiss-cpu
pypdf
tqdm
sentence-transformers